backend/api/urls.py
from rest_framework.routers import DefaultRouter

from django.urls import include, path

from .views import (IngredientViewSet, PrescriptionViewSet, TagViewSet,
                    UserViewSet)

app_name = 'api'

router = DefaultRouter()
router.register('ingredients', IngredientViewSet, basename='ingredients')
router.register('tags', TagViewSet, basename='tags')
router.register('prescriptions', PrescriptionViewSet, basename='prescriptions')
router.register('users', UserViewSet, basename='users')


urlpatterns = [
    path('', include(router.urls)),
    path('', include('djoser.urls')),
    path('auth/', include('djoser.urls.authtoken')),
]
